// types/Breed.tsx
// types/Breed.ts
export interface Breed {
    adaptabilityLevel: number;
    affectionWithFamily: number;
    barkingLevel: number;
    breedGroup: string;
    coatLength: string;
    coatType: string;
    description: string;
    droolingLevel: number;
    energyLevel: number;
    englishName: string;
    exercise: string;
    goodWithOtherDogs: number;
    goodWithYoungChildren: number;
    grooming: string;
    groomingLevel: number;
    guardProtectiveInstinct: number;
    health: string;
    height: string;
    id: number;
    koreanName: string;
    lifeExpectancy: string;
    needsMentalStimulation: number;
    nutrition: string;
    opennessToStrangers: number;
    origin: string;
    playfulnessLevel: number;
    popularity: number;
    sheddingLevel: number;
    size: string;
    temperament: string;
    theme?: string;
    trainabilityLevel: number;
    training: string;
    weight: string;
  }
  